requirements.txt
# 文本处理和分词
jieba>=0.42.1
# 模糊匹配
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
# 语义相似度计算
sentence-transformers>=2.2.0
# 向量操作和存储
numpy>=1.24.0
faiss-cpu>=1.7.4
# 进度条显示
tqdm>=4.65.0
# 配置文件处理
pyyaml>=6.0
# 数据处理和Excel输出
pandas>=2.0.0
openpyxl>=3.1.0